move mount to mount_linux and mount_other
