+++ crates/shim/src/lib.rs
@@ -55,7 +55,14 @@ pub mod cgroup;
 pub mod event;
 pub mod logger;
 pub mod monitor;
-pub mod mount;
+#[cfg(target_os = "linux")]
+pub mod mount_linux;
+#[cfg(not(target_os = "linux"))]
+pub mod mount_other;
+#[cfg(target_os = "linux")]
+pub use mount_linux as mount;
+#[cfg(not(target_os = "linux"))]
+pub use mount_other as mount;
 mod reap;
 #[cfg(not(feature = "async"))]
 pub mod synchronous;
